requirements.txt
# Core dependencies
streamlit>=1.28.0
anthropic>=0.8.0
openai>=1.0.0
qdrant-client>=1.7.0

# Google Cloud integration
google-cloud-storage>=2.10.0
google-api-python-client>=2.100.0
google-auth>=2.23.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Production
gunicorn>=21.2.0
